edit, create entry modal aligned
