+++ client/src/pages/Entries/EditEntry.js
@@ -31,7 +31,7 @@ const theme = createTheme({
   },
 });
 export default function EditEntry({ open, onClose, entryId }) {
-  const isNonMobile = useMediaQuery("(min-width: 1000px)");
+  const isMobile = useMediaQuery("(max-width: 1000px)");
   const [personName, setPersonName] = useState("");
   const [city, setCity] = useState("");
   const [amount, setAmount] = useState("");
@@ -86,7 +86,9 @@ export default function EditEntry({ open, onClose, entryId }) {
             Edit{" "}
           </DialogTitle>
           <DialogContent>
-            <form style={{ paddingTop: 2 }}>
+            <form
+              style={{ paddingTop: 2, width: isMobile ? "250px" : "300px" }}
+            >
               <div
                 style={{
                   display: "flex",
@@ -173,14 +175,14 @@ export default function EditEntry({ open, onClose, entryId }) {
               <br />
               <div
                 style={{
-                  width: "300px",
+                  width: "100%",
                   margin: "5px",
                 }}
               >
                 <label
                   htmlFor="demo-controlled-radio-buttons-group"
                   style={{
-                    width: "300px",
+                    width: "100%",
                     margin: "5px",
                     fontFamily: "Poppins",
                     fontSize: "13px",
